add fake data and test it in workflow
