+++ fitlins/workflows/tests/test_workflow.py
@@ -1,0 +1,19 @@
+from copy import deepcopy
+
+from ..base import init_fitlins_wf
+from ...utils import config
+
+def test_init_fitlins_wf(tmp_path, bids_dir, bids_dset, sample_model):
+    _, database_path = bids_dset
+    out_dir = bids_dir / 'derivatives' / 'fitlins'
+    analysis_level = 'participant'
+    space = "T1w"
+    model = str(sample_model)
+    desc = "preproc"
+
+    run_fitlins = init_fitlins_wf(database_path, out_dir, analysis_level,
+                                  space, model=model, base_dir=tmp_path, desc=desc)
+    run_fitlins.config = deepcopy(config.get_fitlins_config()._sections)
+    run_fitlins.config['execution']['crashdump_dir'] = tmp_path
+    run_fitlins.config['execution']['crashfile_format'] = 'txt'
+    run_fitlins.run()
